+++ influxdb2/client/write_api.py
@@ -202,7 +202,7 @@ class WriteApi(AbstractClient):
 
     def _post_write(self, _async_req, bucket, org, body, precision):
         return self._write_service.post_write(org=org, bucket=bucket, body=body, precision=precision,
-                                              async_req=_async_req)
+                                              async_req=_async_req, content_encoding="utf-8")
 
     def _retryable(self, data: str, delay: timedelta):